rearrange for better unit testing and make way for udp/tcp sessions
